+++ packages/esm-amd-loader/test/run-wct.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 if [ "$WCT_SAUCE" = "true" ]; then
-  npx wct --plugin local --plugin sauce
+  npx wct --plugin sauce
 else
   npx wct --plugin local
 fi