+++ packages/gen-typescript-declarations/src/test/goldens/polymer-2/lib/elements/dom-repeat.d.ts
@@ -41,26 +41,28 @@ declare namespace Polymer {
    *
    *   </template>
    *
-   *   <script>
-   *     class EmployeeList extends Polymer.Element {
-   *       static get is() { return 'employee-list'; }
-   *       static get properties() {
-   *         return {
-   *           employees: {
-   *             value() {
-   *               return [
-   *                 {first: 'Bob', last: 'Smith'},
-   *                 {first: 'Sally', last: 'Johnson'},
-   *                 ...
-   *               ];
-   *             }
-   *           }
-   *         };
-   *       }
-   *     }
-   *   < /script>
-   *
    * </dom-module>
+   * ```
+   *
+   * With the following custom element definition:
+   *
+   * ```js
+   * class EmployeeList extends Polymer.Element {
+   *   static get is() { return 'employee-list'; }
+   *   static get properties() {
+   *     return {
+   *       employees: {
+   *         value() {
+   *           return [
+   *             {first: 'Bob', last: 'Smith'},
+   *             {first: 'Sally', last: 'Johnson'},
+   *             ...
+   *           ];
+   *         }
+   *       }
+   *     };
+   *   }
+   * }
    * ```
    *
    * Notifications for changes to items sub-properties will be forwarded to template
@@ -208,8 +210,8 @@ declare namespace Polymer {
      */
     targetFramerate: number|null|undefined;
     readonly _targetFrameTime: number|null|undefined;
-    disconnectedCallback(): void;
     connectedCallback(): void;
+    disconnectedCallback(): void;
 
     /**
      * Forces the element to render its content. Normally rendering is